test(plan): add empty prompt warning test and split monolith test file (#211)
- Add test for CSA step empty prompt warning branch
- Split plan_cmd_tests.rs (933 lines) into two modules under 800 limit:
  plan_cmd_tests.rs (core tests) and plan_cmd_override_tests.rs (override,
  heartbeat, output forwarding, and empty prompt warning tests)
